feat: Add rules and validator to static method params
