--- newrelic-utils/validator/index.js
@@ -1,5 +1,0 @@
-import RuleNameSpace from './rule';
-import ValidatorNameSpace from './validator';
-
-export const Rule = RuleNameSpace;
-export const Validator = ValidatorNameSpace;
